+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.10.1] - 2026-01-03
 ### Added 
 - CONFIG_EXCHANGE_OPTION constant
+- `is_symbol` to symbol_util
 
 ## [1.10.0] - 2026-01-02
 ### Added